+++ CHANGELOG.md
@@ -6,6 +6,55 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Status
+
+- Development.
+
+## (0.0.9) V 11-09-23
+
+### Added
+
+- Put word method in Board class.
+- Test for Put word method in Board class.
+- Board and menu in CLI.
+- Dockerfile.
+
+### Changed
+
+ - Nothing at the moment.
+
+ ### In Progress
+
+ - Validate Word in Scrabble Game Class.
+ - Test for Validate Word in Scrabble Game Class.
+
+### Issues
+
+- DockerFile Error.
+
+
+
+
+## Status
+
+- Development.
+
+## (0.0.8) V 06-09-23
+
+### Added
+
+- Hierarchy of Players Turns.
+- Test for Hierarchy of Players Turns.
+
+### Changed
+
+ - Nothing at the moment.
+
+### Issues
+
+- Nothing at the moment.
+
+
 
 
 ## Status